yolo5/requirements.txt
flask
pyyaml
loguru
boto3
pymongo
python-dotenv
numpy
ultralytics